add colors and rename files
