+++ src/components/FormContactUs/FormContactUs.styles.ts
@@ -1,5 +1,7 @@
 import styled from 'styled-components'
 
+import { colors } from 'styles/colors'
+
 export const FormWrapper = styled.form`
   display: flex;
   flex-direction: column;
@@ -7,17 +9,17 @@ export const FormWrapper = styled.form`
 `
 export const Error = styled.span`
   font-size: 0.75rem;
-  color: red;
+  color: ${colors.red[500]};
   margin-bottom: 0.25rem;
 `
-export const TitleForm = styled.h1`
+export const TitleForm = styled.h2`
   text-align: center;
   margin: 1rem 0rem;
 `
 
 export const ContainerForm = styled.div`
   padding: 0.75rem;
-  background-color: #eceaea;
+  background-color: ${colors.gray[100]};
   border-radius: 1rem;
   width: 100%;
 `
