+++ src/pages/ContactUs/ContactUs.tsx
@@ -3,7 +3,7 @@ import { PopupButton } from 'react-calendly'
 import './calendly.css'
 import useIntlMessages from 'hooks/useIntlMessages'
 import Button from 'ui/Button/Button'
-import FormContacUs from 'components/FormContactUs/Form'
+import FormContacUs from 'components/FormContactUs/FormContactUs'
 import {
   ContactUsHero,
   HeroWrapper,
